requirements.txt
streamlit
ultralytics==8.3.63
opencv-python-headless 
numpy
Pillow